improve app.js

+++ app.js
@@ -15,13 +15,34 @@ if (fs.existsSync(SESSION_FILE_PATH)) {
     sessionCfg = require(SESSION_FILE_PATH);
 }
 
-const client = new Client({ puppeteer: { headless: true, args: ['--no-sandbox'] }, session: sessionCfg });
+let client = new Client({ puppeteer: { headless: true, args: ['--no-sandbox'] }, session: sessionCfg });
 const routeMain = new RouteMain();
 routeMain.setBotClient(client);
 routeMain.setQRModule(QRCode);
 
+client.on('disconnected', (reason) => {
+    console.log('Client was logged out', reason);
+
+    if (fs.existsSync(SESSION_FILE_PATH)) {
+        // remove current session file
+        fs.unlinkSync(SESSION_FILE_PATH);
+        sessionCfg = null;
+    }
+
+    routeMain.isClientReady(false);
+
+    //destroy client instant 
+    client.destroy().then(function () {
+        console.log('Client is shutdown..');
+        //start whatsapp client engine 
+        client.initialize().then(function () {
+            console.log('Fresh copy is ready..');
+        });
+    });
+});
+
 client.on('authenticated', (session) => {
-    console.log('AUTHENTICATED', session);
+    console.log('login sucess..');
     sessionCfg = session;
     fs.writeFile(SESSION_FILE_PATH, JSON.stringify(session), function (err) {
         if (err) {
@@ -36,27 +57,28 @@ client.on('auth_failure', msg => {
 
     // remove current session file
     fs.unlinkSync(SESSION_FILE_PATH);
+    sessionCfg = null;
 
     //destroy client instant 
     client.destroy().then(function () {
-        /*
-            we are using docker with restart:always, exit the 
-            system to make container reboot, so we get a new fresh copy
-            remember, there is a downtime.
-        */
-        process.exit(1);
+        console.log('Client is shutdown..');
+        //start whatsapp client engine 
+        client.initialize().then(function () {
+            console.log('Fresh copy is ready..');
+        });
     });
 });
 
 client.on('qr', (qr) => {
     // Generate and scan this code with your phone
     routeMain.setQRCode(qr);
-    console.log("QR is ready");
+    console.log("qrcode is ready..");
 });
 
 client.on('ready', () => {
     routeMain.setQRCode(null);
-    console.log('Client is ready!');
+    routeMain.isClientReady(true);
+    console.log('Client is ready..');
 });
 
 client.on('message', async msg => {
@@ -72,22 +94,29 @@ client.on('message', async msg => {
 //define route request
 app.get('/', routeMain.handleIndex.bind(routeMain));
 app.get('/device', routeMain.handleDevice.bind(routeMain));
+app.post('/', routeMain.handleIndex.bind(routeMain));
+app.post('/device', routeMain.handleDevice.bind(routeMain));
 app.get('/qr/scan', routeMain.handleScanner.bind(routeMain));
 
 //reset will remove session so we need to scan new qrcode
 app.get('/api/reset', function (req, res) {
+    if (fs.existsSync(SESSION_FILE_PATH) == false || routeMain.clientReady == false) {
+        res.setHeader('Content-Type', 'Application/Json');
+        res.send(JSON.stringify({
+            info: false,
+            status: 'Reset a non ready client is illegal.'
+        }));
+
+        return;
+    }
+
+    console.log('Shuting down the client..');
     // remove current session file
     fs.unlinkSync(SESSION_FILE_PATH);
+    sessionCfg = null;
 
-    //destroy client instant 
-    client.destroy().then(function () {
-        /*
-            we are using docker with restart:always, exit the 
-            system to make container reboot, so we get a new fresh copy
-            remember, there is a downtime.
-        */
-        process.exit(1);
-    });
+    //the best solution for reseting client is reset the container it self
+    process.exit(1);
 });
 
 //cek server status, to check uptime after downtime
@@ -100,7 +129,9 @@ app.get('/api/health', function (req, res) {
 });
 
 //start whatsapp client engine 
-client.initialize();
+client.initialize().catch(function (err) {
+    console.error(err);
+});
 
 //start http server 
-app.listen(port, host, () => console.log(`listening at http://${host}:${port}`))
+app.listen(port, host, () => console.log(`listening at http://${host}:${port}`));
